% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevo.R
\docType{data}
\name{demographicData}
\alias{demographicData}
\title{Draws for observed heterogeneity in Nevo's cereal example.}
\format{Draws for observed heterogeneity for each demographic.
\describe{
  \item{cdid}{market identifier,}
  \item{draws_}{20 draws differing across markets.}
}}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/26803/SOF9FW&version=1.0}
}
\usage{
demographicData
}
\description{
Draws for observed heterogeneity in Nevo's cereal example.
}
\keyword{datasets}
