% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticitiesBLP.R
\name{elasticitiesBLP}
\alias{elasticitiesBLP}
\title{Conveniently extract elasticities from result of estimateBLP}
\usage{
elasticitiesBLP(blp, var = names(blp$elasticities)[[1]], market = 1,
  products = NULL, round.digits = 3)
}
\arguments{
\item{blp}{The result from estimateBLP}

\item{var}{The variable name of the elasticity to be estimated}

\item{market}{Number or name of the market for which elasticities shall be computed}

\item{products}{A character vector or NULL (default). If not NULL only return elasticities for the specified products. Looks up product names in the column specified by \code{prodid} in the call to \code{estimateBLP}}

\item{round.digits}{If not NULL the number of digits to which elasticities shall be rounded (DEFAULT=3)}
}
\description{
Currently only works for variables whose name was added to the argument \code{postEstimation.control$elasticities} in the call to \code{estimateBLP}
}
