% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevo.R
\docType{data}
\name{originalDraws}
\alias{originalDraws}
\title{Draws for unobserved heterogeneity in Nevo's cereal example.}
\format{Draws for unobserved heterogeneity for each random coefficient.
\describe{
  \item{cdid}{market identifier,}
  \item{draws_}{20 draws differing across markets.}
}}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/26803/SOF9FW&version=1.0}
}
\usage{
originalDraws
}
\description{
Draws for unobserved heterogeneity in Nevo's cereal example.
}
\keyword{datasets}
