% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cereal.R
\docType{data}
\name{productData_cereal}
\alias{productData_cereal}
\title{Product data of Nevo's cereal example.}
\format{A data frame with product data of 24 cereals in each of 94 markets.
\describe{
  \item{share}{cereals market share,}
  \item{price}{cereals price,}
  \item{const}{constant,}
  \item{sugar}{cereals sugar,}
  \item{mushy}{cereals mushy,}
  \item{cdid}{market identifier,}
  \item{product_id}{uniquely identifies a product in a market,}
  \item{productdummy}{uniquely identifies a product in a market,}
  \item{IV1}{1. instrument,}
  \item{IV2}{2. instrument,}
 \item{IV3}{3. instrument,}
 \item{IV4}{4. instrument,}
 \item{IV5}{5. instrument,}
 \item{IV6}{6. instrument,}
 \item{IV7}{7. instrument,}
 \item{IV8}{8. instrument,}
 \item{IV9}{9. instrument,}
 \item{IV10}{10. instrument,}
 \item{IV11}{11. instrument,}
 \item{IV12}{12. instrument,}
 \item{IV13}{13. instrument,}
 \item{IV14}{14. instrument,}
 \item{IV15}{15. instrument,}
 \item{IV16}{16. instrument,}
 \item{IV17}{17. instrument,}
 \item{IV18}{18. instrument,}
 \item{IV19}{19. instrument,}
 \item{IV20}{20. instrument}
   }}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/26803/SOF9FW&version=1.0}
}
\usage{
productData_cereal
}
\description{
Product data of Nevo's cereal example.
}
\keyword{datasets}
