% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cereal.R
\docType{data}
\name{theta_guesses_cereal}
\alias{theta_guesses_cereal}
\title{Parameter starting guesses for Nevo's cereal example.}
\format{A matrix with 4 random coefficients (rows) and columns for 4 demographics and one unobserved heterogeneity column (5 cols in total).}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/26803/SOF9FW&version=1.0}
}
\usage{
theta_guesses_cereal
}
\description{
Parameter starting guesses for Nevo's cereal example.
}
\keyword{datasets}
