% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nevo.R
\docType{data}
\name{w_guesses}
\alias{w_guesses}
\title{Mean utility starting guesses for Nevo's cereal example.}
\format{A numeric vector of 2256 values.}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/26803/SOF9FW&version=1.0}
}
\usage{
w_guesses
}
\description{
Mean utility starting guesses for Nevo's cereal example.
}
\keyword{datasets}
